+++ ta4j-core/src/test/java/org/ta4j/core/analysis/ReturnsTest.java
@@ -9,6 +9,8 @@ import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.PrecisionNum;
 
+import java.math.BigDecimal;
+import java.math.MathContext;
 import java.util.function.Function;
 
 import static org.junit.Assert.assertEquals;
@@ -56,6 +58,52 @@ public class ReturnsTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
         assertNumEquals(1d/5, strategyReturns.getValue(4));
         assertNumEquals(0, strategyReturns.getValue(5));
         assertNumEquals(1 - (20d/3), strategyReturns.getValue(6));
+    }
+
+    @Test
+    public void returnsShortLeverageArithmetic() {
+        TimeSeries sampleTimeSeries = new MockTimeSeries(numFunction,10, 15, 5, 1, 1);
+        TradingRecord tradingRecord = new BaseTradingRecord(
+                Order.sellAt(0, sampleTimeSeries), Order.buyAt(4, sampleTimeSeries));
+
+        Returns strategyReturns = new Returns(sampleTimeSeries, tradingRecord, Returns.ReturnType.ARITHMETIC);
+
+        assertNumEquals(NaN.NaN, strategyReturns.getValue(0));
+        assertNumEquals(-0.5, strategyReturns.getValue(1));
+        assertNumEquals(2, strategyReturns.getValue(2));
+        assertNumEquals(4.0/15, strategyReturns.getValue(3));
+        assertNumEquals(0, strategyReturns.getValue(4));
+
+        Num one = numOf(1);
+        Num totalReturn = one.plus(strategyReturns.getValue(1)).multipliedBy(
+                one.plus(strategyReturns.getValue(2))).multipliedBy(
+                one.plus(strategyReturns.getValue(3))).multipliedBy(
+                one.plus(strategyReturns.getValue(4)));
+
+        assertNumEquals(2.0 - 1.0/10, totalReturn);
+    }
+
+    @Test
+    public void returnsShortLeverageLog() {
+        TimeSeries sampleTimeSeries = new MockTimeSeries(numFunction,10, 15, 5, 1, 1);
+        TradingRecord tradingRecord = new BaseTradingRecord(
+                Order.sellAt(0, sampleTimeSeries), Order.buyAt(4, sampleTimeSeries));
+
+        Returns strategyReturns = new Returns(sampleTimeSeries, tradingRecord, Returns.ReturnType.LOG);
+
+        assertNumEquals(NaN.NaN, strategyReturns.getValue(0));
+        assertNumEquals(-Math.log(15.0/10), strategyReturns.getValue(1));
+        assertNumEquals(-Math.log(5.0/15), strategyReturns.getValue(2));
+        assertNumEquals(-Math.log(1.0/5), strategyReturns.getValue(3));
+        assertNumEquals(0, strategyReturns.getValue(4));
+
+        Num one = numOf(1);
+        Num totalReturn = strategyReturns.getValue(1)
+                .plus(strategyReturns.getValue(2))
+                .plus(strategyReturns.getValue(3))
+                .plus(strategyReturns.getValue(4));
+
+        assertNumEquals(-Math.log(1.0/10), totalReturn);
     }
 
     @Test
@@ -115,6 +163,6 @@ public class ReturnsTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
         assertNumEquals(arithPrecision, PrecisionNum.valueOf(1.1).dividedBy(PrecisionNum.valueOf(1.2)).minus(PrecisionNum.valueOf(1)));
 
         assertNumEquals(logDouble, DoubleNum.valueOf(-0.08701137698962969));
-        assertNumEquals(logPrecision, PrecisionNum.valueOf(-0.08701137698962981));
+        assertNumEquals(logPrecision, PrecisionNum.valueOf("-0.087011376989629766167765901873746"));
     }
 }